Ft: Build Dto

+++ Controllers/AccountController.cs
@@ -1,5 +1,7 @@
 ﻿using Booking_Api.Contracts;
+using Booking_Api.DTOs.Accounts;
 using Booking_Api.Models;
+using Booking_Api.Utilities.Handler;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Booking_Api.Controllers
@@ -25,7 +27,9 @@ namespace Booking_Api.Controllers
                 return NotFound("Data Not Found");
             }
 
-            return Ok(accounts);
+            var data = accounts.Select(x => (AccountDto)x);
+
+            return Ok(data);
         }
 
         // Get: api/Account/guid
@@ -37,38 +41,53 @@ namespace Booking_Api.Controllers
             {
                 return NotFound("Account Not found");
             }
-            return Ok(account);
+            return Ok((AccountDto)account);
         }
 
         // Post: api/Account
         [HttpPost]
-        public IActionResult Create(Account account)
+        public IActionResult Create(CreateAccountDto createAccountDto)
         {
-            var createdAccount = _accountRepository.Create(account);
+            Account toCreate = createAccountDto;
+            toCreate.Password = HashingHandler.HashPassword(createAccountDto.Password);
+
+            var createdAccount = _accountRepository.Create(createAccountDto);
             if (createdAccount is null)
             {
                 return BadRequest("Not Created Account. Try Again!");
             }
-            return Ok(createdAccount);
+            return Ok((AccountDto)createdAccount);
         }
 
         // Put: api/Account
         [HttpPut]
-        public IActionResult Update(Account account)
+        public IActionResult Update(AccountDto accountDto)
         {
+            var account = _accountRepository.GetById(accountDto.Guid);
+            if (account is null)
+            {
+                return NotFound("Id Not Found");
+            }
+            Account toUpdate = accountDto;
+            toUpdate.CreatedDate = account.CreatedDate;
+            toUpdate.Password = HashingHandler.HashPassword(accountDto.Password);
             var updatedAccount = _accountRepository.Update(account);
             if (!updatedAccount)
             {
                 return BadRequest("Not Updated Account. Try Again!");
             }
-            return Ok(updatedAccount);
+            return Ok("Data has been updated successfully");
         }
 
         // Delete: api/Account
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var account = new Account() { Guid = guid };
+            var account = _accountRepository.GetById(guid);
+            if (account is null)
+            {
+                return NotFound("Id Not Found");
+            }
             var deletedAccount = _accountRepository.Delete(account);
             if (!deletedAccount)
             {
