+++ Controllers/EmployeController.cs
@@ -1,5 +1,7 @@
 ﻿using Booking_Api.Contracts;
+using Booking_Api.DTOs.Employees;
 using Booking_Api.Models;
+using Booking_Api.Repositories;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Booking_Api.Controllers
@@ -25,7 +27,9 @@ namespace Booking_Api.Controllers
                 return NotFound("Data Not Found");
             }
 
-            return Ok(employes);
+            var data = employes.Select(x => (EmployeeDto)x);
+
+            return Ok(data);
         }
 
         // Get: api/Employe/guid
@@ -37,38 +41,50 @@ namespace Booking_Api.Controllers
             {
                 return NotFound("Employe Not found");
             }
-            return Ok(employe);
+            return Ok((EmployeeDto)employe);
         }
 
         // Post: api/Employe
         [HttpPost]
-        public IActionResult Create(Employe employe)
+        public IActionResult Create(CreateEmployeeDto createEmployeeDto)
         {
-            var createdEmploye = _employeRepository.Create(employe);
+            var createdEmploye = _employeRepository.Create(createEmployeeDto);
             if (createdEmploye is null)
             {
                 return BadRequest("Not Created Employe. Try Again!");
             }
-            return Ok(createdEmploye);
+            return Ok((EmployeeDto)createdEmploye);
         }
 
         // Put: api/Employe
         [HttpPut]
-        public IActionResult Update(Employe employe)
+        public IActionResult Update(EmployeeDto employeeDto)
         {
+            var employe = _employeRepository.GetById(employeeDto.Guid);
+            if (employe is null)
+            {
+                return NotFound("Id Not Found");
+            }
+            Employe toUpdate = employeeDto;
+            toUpdate.CreatedDate = employe.CreatedDate;
+
             var updatedEmploye = _employeRepository.Update(employe);
             if (!updatedEmploye)
             {
                 return BadRequest("Not Updated Employe. Try Again!");
             }
-            return Ok(updatedEmploye);
+            return Ok("Data has been updated successfully");
         }
 
         // Delete: api/Employe
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var employe = new Employe() { Guid = guid };
+            var employe = _employeRepository.GetById(guid);
+            if (employe is null)
+            {
+                return NotFound("Id Not Found");
+            }
             var deletedEmploye = _employeRepository.Delete(employe);
             if (!deletedEmploye)
             {
