+++ Controllers/RoleController.cs
@@ -1,4 +1,6 @@
 ﻿using Booking_Api.Contracts;
+using Booking_Api.DTOs.Roles;
+using Booking_Api.DTOs.Rooms;
 using Booking_Api.Models;
 using Microsoft.AspNetCore.Mvc;
 
@@ -25,7 +27,9 @@ namespace Booking_Api.Controllers
                 return NotFound("Data Not Found");
             }
 
-            return Ok(roles);
+            var data = roles.Select(x => (RoleDto)x);
+
+            return Ok(data);
         }
 
         // Get: api/Role/guid
@@ -37,38 +41,50 @@ namespace Booking_Api.Controllers
             {
                 return NotFound("Role Not found");
             }
-            return Ok(role);
+            return Ok((RoleDto)role);
         }
 
         // Post: api/Role
         [HttpPost]
-        public IActionResult Create(Role role)
+        public IActionResult Create(CreateRoleDto createRoleDto)
         {
-            var createdRole = _roleRepository.Create(role);
+            var createdRole = _roleRepository.Create(createRoleDto);
             if (createdRole is null)
             {
                 return BadRequest("Not Created Role. Try Again!");
             }
-            return Ok(createdRole);
+            return Ok((RoleDto)createdRole);
         }
 
         // Put: api/Role
         [HttpPut]
-        public IActionResult Update(Role role)
+        public IActionResult Update(RoleDto roleDto)
         {
+            var role = _roleRepository.GetById(roleDto.Guid);
+            if (role is null)
+            {
+                return NotFound("Id Not Found");
+            }
+            Role toUpdate = roleDto;
+            toUpdate.CreatedDate = role.CreatedDate;
+
             var updatedRole = _roleRepository.Update(role);
             if (!updatedRole)
             {
                 return BadRequest("Not Updated Role. Try Again!");
             }
-            return Ok(updatedRole);
+            return Ok("Data has been updated successfully");
         }
 
         // Delete: api/Role
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var role = new Role() { Guid = guid };
+            var role = _roleRepository.GetById(guid);
+            if (role is null)
+            {
+                return NotFound("Id Not Found");
+            }
             var deletedRole = _roleRepository.Delete(role);
             if (!deletedRole)
             {
