+++ Controllers/RoomController.cs
@@ -1,4 +1,5 @@
 ﻿using Booking_Api.Contracts;
+using Booking_Api.DTOs.Rooms;
 using Booking_Api.Models;
 using Microsoft.AspNetCore.Mvc;
 
@@ -24,8 +25,9 @@ namespace Booking_Api.Controllers
             {
                 return NotFound("Data Not Found");
             }
+            var data = rooms.Select(x => (RoomDto)x);
 
-            return Ok(rooms);
+            return Ok(data);
         }
 
         // GET: api/Room/guid
@@ -37,38 +39,51 @@ namespace Booking_Api.Controllers
             {
                 return NotFound("Room Not found");
             }
-            return Ok(room);
+            return Ok((RoomDto)room);
         }
 
         // POST: api/Room
         [HttpPost]
-        public IActionResult Create(Room room)
+        public IActionResult Create(CreateRoomDto createRoomDto)
         {
-            var createdRoom = _roomRepository.Create(room);
+            var createdRoom = _roomRepository.Create(createRoomDto);
             if (createdRoom is null)
             {
                 return BadRequest("Not Created Room. Try Again!");
             }
-            return Ok(createdRoom);
+            return Ok((RoomDto)createdRoom);
         }
 
         // PUT: api/Room
         [HttpPut]
-        public IActionResult Update(Room room)
+        public IActionResult Update(RoomDto roomDto)
         {
+            var room = _roomRepository.GetById(roomDto.Guid);
+            if (room is null)
+            {
+                return NotFound("Id Not Found");
+            }
+
+            Room toUpdate = roomDto;
+            toUpdate.CreatedDate = room.CreatedDate;
+
             var updatedRoom = _roomRepository.Update(room);
             if (!updatedRoom)
             {
                 return BadRequest("Not Updated Room. Try Again!");
             }
-            return Ok(updatedRoom);
+            return Ok("Data has been updated successfully");
         }
 
         // DELETE: api/Room
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-            var room = new Room() { Guid = guid };
+            var room = _roomRepository.GetById(guid);
+            if (room is null)
+            {
+                return NotFound("Id Not Found");
+            }
             var deletedRoom = _roomRepository.Delete(room);
             if (!deletedRoom)
             {
